Finish clean up.
