+++ reflutter_test/test/api_integration_tests.dart
@@ -15,20 +15,6 @@ void main() async {
     return new Future<Response>.sync(() => resp);
   });
 
-  test("Simple serialization", () async {
-    var jBody = json.encode(new HealthResponse("OK"));
-    expect(true, true);
-
-    var sBody = new HealthResponse.fromJson(json.decode(jBody));
-    expect(sBody.value, "OK");
-  });
-
-  test("List serialization", () async {
-    var list = new List<HealthResponse>.filled(5, new HealthResponse("OK"));
-    var jBody = json.encode(list);
-    var obj = json.decode(jBody);
-  });
-
   test("Test GET call", () async {
     final api = new TestApi(mockclient, "/", {});
     var resp = await api.healthcheck();
