--- reflutter_test/test/serialization.dart
@@ -1,9 +1,0 @@
-library serialization;
-
-import 'package:test/test.dart';
-
-void main() {
-  test("Always true", () {
-    expect(true, true);
-  });
-}
